floating math floors

+++ 0x00-python_variable_annotations/2-floor.py
@@ -2,6 +2,6 @@
 """Basic annotations - floor"""
 import math
 
-def floor(n: int) -> int:
+def floor(n: float) -> int:
     """Returs floor"""
     return math.floor(n)
